requirements.txt
extruct
scrapy
fake-useragent
unidecode
python-slugify
pandas